Preliminary Level Editor
Using Tiled Map Editor you can create the game level and you can have it output the data in an XML file. The only data that is really important is tile positions so you can just copy and paste that into a file with a .map extension. There are some sample files to go along with this commit. input1 is the tile sheet and testlevel.map is an example .map file with dimensions of the level on the first line. Make sure to change Program.cs to use Game3 instead of Game1 to run this test.