Tile Engine is complete I believe. It draws tiles from a tile sheet to the screen using a .map file with the data of which square when. There is a sample tile sheet that draws that sample level I made. It is with the wrong tile sheet, but it works. There is more detailed information on how exactly to use it on the Game3.cs file.
You can draw the maps with Tiled Map Editor, change the properties to write the tile data in CSV in the properties and then open up the file in notepad or VS or w/e and copy and paste the map data into your .map file for a way to draw with the tiles visually and have an easy way to convert the visual drawer into your .map file.

Same as previous commit but with bug fixes.