New features added!
TileMap can now handle some characters as well as numbers.
S: sets up the players starting position.
C: sets up a checkpoint within a level.
X: sets up the level's end square.

Allows for easy placement of those items within the game as they are just included within the .map files.

All edited files are modified to handle placing and setting up those locations as well as checking when the player hits them.

As of now level changes are handled on a mod; at the end of the level cycle, it starts back at the beginning.

I added another level, this one is quite a bit bigger and also is a bit more challenging with its jumps.