Player.cs
Fixed the movement so that the Xbox 360 D-Pad will now work.

Screens/GameOverScreen.cs
Added a game over screen based off PauseMenuScreen.cs.  When all lives are lost, this will pop up.  If the user selects retry, the level reloads with 3 lives.  However, they can also quit to return to the main menu from here.

Screens/GameplayScreen.cs
Bunch of changes.  First I modified the levelIndex value to pull from the GameOptions class from the OptionsMenuScreen and subtracted 2 from it (so 1 is -1 like it use to be).  This way if the user selects to go to level 2 in the Options Menu/Level Select screen, the game will start in the second level.  Added in the pause function so that the game can now be paused.  Changed how input is found slightly.  Gives added advantage so that if someone is using an Xbox 360 controller, if it's disconnected the game will pause automatically.  Added in a game over function that pops up when player is out of lives.  If retry is selected, level reloads.  If main menu is selected, return to the main menu.  Drew in a life counter in the bottom left corner of the screen.

Screens/MainMenuScreen.cs
Changed the Options menu option into a Level Select menu option.

Screens/OptionsMenuScreen.cs
Removed almost everything that was originally in it.  Added in a level select menu function.  Allows users to select between level 1 and 2 for now, can easily be expanded for additional levels.  Creates a GameOptions class so I can send values from OptionsMenuScreen to all the other screens, mostly so the level select function works properly.

Sprites/Player/lifeIcon.png
Added an icon for the HUD's life counter.