The Xbox 360 Enabled and Screen Size Change Update
File Movement:
Moved background.png, blank.png and gradient.png into Backgrounds foulder.  Moved menufont.spritefont and gamefont.spritefont into the Fonts folder.

Game.cs
Fixed the content load for graident to reflect it's new folder location.  Added in an autodetect setting in order to change the size of the screen.  The game now supports the resolutions of 1024x768, 640x480 and 1280x720 along with it's original 800x600.

ScreenManager/ScreenManager.cs
Fixed the content load for menufont and blank to reflect it's new folder location.

Screens/BackgroundScreen.cs
Fixed the content load for background to reflect it's new folder location.

Screens/GameplayScreen.cs
Modified the HUD code to now have different versions for Xbox 360 and Windows.  Mostly looks the same, but titleSafeArea.Bottom actually shows up on the Xbox 360, but not on Windows.

Screens/MainMenuScreen.cs
Changed the message so that on smaller screen resolutions full message is shown.

Screens/MessageBoxScreen.cs
Changed the output messages to reflect what system it's on.  On an Xbox 360, only Xbox 360 controlls will be show.  Windows will still display both.  Fixed the content load for graident to reflect it's new folder location.

Screens/PauseMenuScreen.cs
Changed the message so that on smaller screen resolutions full message is shown.