The (Sorta) Windows Phone 7 Compatible Update
Although it doesn't work, I went ahead and added in the Windows Phone 7 code and screens.  It shows how I allow for touch screen and accelerometer input as well as new menu, game over and pause screens.  Also modified the file structure, placing files only needed in for Windows and Xbox into their own folder, and files only needed for Windows Phone into their own folders.  Everything still works perfectly fine.  The only difference is that now if you want to look at the code I did, it's there.  The control code and menus are good, but for some reason on a phone TileMap fails to load.  That's the only reason it doesn't work.  If it did load though, the thing would be flawless.