The Quit Menu Text Change Update
Added in a nice feature where if a controller is not connected on the PC version, it won't show the Xbox control options when you try to quit.  If a controller is present, then both keyboard and controller options are shown.  As well, if a controller is disconnected and the player resumes on a keyboard, the game will resume as if a controller has never been present.  Quitting now would also show the keyboard only controls.  Not really needed or critical, but it's a nice touch.

In more file moving, I put thumbstick in it's own folder and changed it's call to reflect that.