Xbox 360 and Windows Phone Project Commit
Created Xbox 360 and Windows Phone copies of the main project and renamed the main project "Windows".  Did some HUD readjustment for the Windows and Xbox versions.  Set the Windows Phone version to use the same as the Windows version, but I got no idea how that looks.  If I ever figure out why the problem I have with it happens, I might one day go back in and fix it.

But, it's there and it's fully ready to roll on my 360.